* fixed compiler warning concerning comparison of signed and unsigned int @BitSplitter.cpp
git-archimport-id: falk@cs.fau.de--2007-systemoc-intern/SystemC-Examples--mainline--0.6--patch-248

+++ systemoc-jpeg-decoder/BitSplitter.cpp
@@ -144,7 +144,7 @@ void BitSplitter::dumpBuffer(std::ostream &out) const
 //
 void BitSplitter::skipByte(const int num) {
   //std::cerr << " skipByte()" << std::endl;
-  assert(m_bufferNum >= num);
+  assert((int)m_bufferNum >= num);
   m_bufferNum -= num;
   m_readPos = (m_readPos + num) % m_bufferSize;
